Separate into using one hook for data about a single column, and one hook for data about all columns together.
