+++ src/features/import/components/ImportDialog/Configure/Configuration/index.tsx
@@ -1,29 +1,28 @@
-import { CompareArrows } from '@mui/icons-material';
 import { FC } from 'react';
 import { Box, useTheme } from '@mui/material';
 
 import IdConfig from './IdConfig';
-import messageIds from 'features/import/l10n/messageIds';
 import OrgConfig from './OrgConfig';
 import TagConfig from './TagConfig';
-import { UIDataColumn } from 'features/import/hooks/useUIDataColumns';
-import { useMessages } from 'core/i18n';
-import ZUIEmptyState from 'zui/ZUIEmptyState';
 import {
-  Column,
   ColumnKind,
   IDFieldColumn,
   OrgColumn,
   TagColumn,
 } from 'features/import/utils/types';
+import useUIDataColumn, {
+  UIDataColumn,
+} from 'features/import/hooks/useUIDataColumn';
 
 interface ConfigurationProps {
-  uiDataColumn: UIDataColumn<Column> | null;
+  columnIndexBeingConfigured: number;
 }
 
-const Configuration: FC<ConfigurationProps> = ({ uiDataColumn }) => {
-  const messages = useMessages(messageIds);
+const Configuration: FC<ConfigurationProps> = ({
+  columnIndexBeingConfigured,
+}) => {
   const theme = useTheme();
+  const uiDataColumn = useUIDataColumn(columnIndexBeingConfigured);
 
   return (
     <Box
@@ -46,20 +45,6 @@ const Configuration: FC<ConfigurationProps> = ({ uiDataColumn }) => {
         uiDataColumn.originalColumn.kind == ColumnKind.ORGANIZATION && (
           <OrgConfig uiDataColumn={uiDataColumn as UIDataColumn<OrgColumn>} />
         )}
-      {!uiDataColumn && (
-        <Box
-          alignItems="center"
-          display="flex"
-          height="100%"
-          justifyContent="center"
-          sx={{ opacity: '50%' }}
-        >
-          <ZUIEmptyState
-            message={messages.configuration.mapping.emptyStateMessage()}
-            renderIcon={(props) => <CompareArrows {...props} />}
-          />
-        </Box>
-      )}
     </Box>
   );
 };
